Allow module to work with bundled gem
A few things have changed now that the gem is bundled. We don't need to
manage installation anymore, and it's scheduled with the built-in
tkscheduler. Instead of managing a cron job, we'll just turn it on.

Note that I didn't bother to update the Litmus tests because they'll all
fail until the images are updated with the latest release, and I don't
know how to choose the Puppetserver version to install anyway! 🤷